% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\docType{package}
\name{liu-package}
\alias{liu}
\alias{liu-package}
\title{liu: Lightning Index Units}
\description{
LIU implements SQL-like indexes for data frames in R. It allows to index columns of data frames by creating light B+trees with pairs key-row_number. This shifts complexity of searching from linear to logaritmic (O(log n)), which makes few other functions (exp. min, max, count, merge) much faster.
}
\author{
\strong{Maintainer}: Marek Skok \email{marekskok@pw.edu.pl}

}
