% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\name{liu_build}
\alias{liu_build}
\title{Build LIU Index}
\usage{
liu_build(df, column_name)
}
\arguments{
\item{df}{Data frame to be indexed.}

\item{column_name}{Character string specifying the column to index. 
(int, double or string column)}
}
\value{
C LIU index object (pointer).
}
\description{
This function if foundation of LIU package. It builds a high-performance 
SQL-like B+Tree index for a data frame column.
Works only for columns of ints, doubles or strings. Values of given column will be 
put in B+Tree with their row numbers. It ignores NA.
}
\examples{
\dontrun{
df <- data.frame(
id = as.integer(c(1,2,3)),
val = c(3.5,6.7,2.1),
country = c("Poland", "Hungary", "England")
)
idx_int <- liu_build(df, "id")
idx_double <- liu_build(df, "val")
idx_string <- liu_build(df, "country")
}
}
