% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\name{liu_isin}
\alias{liu_isin}
\title{Is In LIU Index}
\usage{
liu_isin(index, keys)
}
\arguments{
\item{index}{A LIU index object (external pointer).}

\item{keys}{Vector of keys (int or double must match LIU index type) to check.}
}
\value{
Vector of logical values.
}
\description{
Checks if given keys are present in LIU index. For NA it returns FALSE.
}
\examples{
\dontrun{
df <- data.frame(
  id = as.integer(c(1,2,3,1)),
  val = c(0.5,6.7,2.1,0.5)
)
idx <- liu_build(df, "val")

logical <- liu_isin(idx, c(0.5,NA,2.1,3.4))
# [1] TRUE FALSE TRUE FALSE

logical <- liu_isin(idx, 6.7)
# [1] TRUE
}
}
