% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\name{liu_min}
\alias{liu_min}
\title{Minimum Key in LIU Index}
\usage{
liu_min(index)
}
\arguments{
\item{index}{A LIU index object (external pointer).}
}
\value{
An integer vector of row indices corresponding to the minimum key.
}
\description{
Search for the smallest key in LIU index and returs their row indices.
}
\examples{
\dontrun{
df <- data.frame(
  id = as.integer(c(1,2,3,1)),
  val = c(0.5,6.7,2.1,0.5),
  country = c("Poland", "Hungary", "England")
)
# Could be be build also on "id" and "country"
idx <- liu_build(df, "val")

# Get row indices for the smallest and largest key in the index
min_rows <- liu_min(idx)
# [1] 1 4
max_rows <- liu_max(idx)
# [1] 2
}
}
