% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\name{liu_search}
\alias{liu_search}
\title{Search for Keys in LIU Index}
\usage{
liu_search(index, key)
}
\arguments{
\item{index}{LIU index object (external pointer).}

\item{key}{Vector of keys (integer, double or character must match LIU index type) to search for.}
}
\value{
Integer vector of row indices where the keys were found. 
Returns an empty vector if none of the keys exist.
}
\description{
Performs a fast lookup in the LIU index to find all row indices associated 
with given vector of keys. It ignores NA in keys vector.
}
\examples{
\dontrun{
df <- data.frame(
  id = as.integer(c(1,2,3,4)),
  val = c(0.5,6.7,2.1,0.5)
)
idx <- liu_build(df, "val")

row_idx <- liu_search(idx, 6.7)
# [1] 2

row_ids <- liu_search(idx, c(6.7, 0.5))
# [1] 2 1 4
}
}
